+++ src/main/java/com/lzt/ssm/blog/service/impl/TagServiceImpl.java
@@ -2,10 +2,12 @@ package com.lzt.ssm.blog.service.impl;
 
 import com.lzt.ssm.blog.entity.Tag;
 import com.lzt.ssm.blog.entity.TagExample;
+import com.lzt.ssm.blog.mapper.ArticleTagRefMapper;
 import com.lzt.ssm.blog.mapper.TagMapper;
 import com.lzt.ssm.blog.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -19,13 +21,22 @@ public class TagServiceImpl implements TagService {
     @Autowired
     private TagMapper tagMapper;
 
+    @Autowired
+    private ArticleTagRefMapper articleTagRefMapper;
+
     @Override
     public void insertEntity(Tag tag) {
         tagMapper.insertSelective(tag);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void deleteEntityById(Integer id) {
+        Integer count = articleTagRefMapper.countArticleByTagId(id);
+        if (count > 0) {
+            throw  new RuntimeException("该标签下存在文章，不可删除");
+        }
+
         tagMapper.deleteByPrimaryKey(id);
     }
 
@@ -41,7 +52,12 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public List<Tag> listEntity() {
-        TagExample tagExample = new TagExample();
-        return tagMapper.selectByExample(tagExample);
+        List<Tag> tagList = tagMapper.selectByExample(null);
+        //填充文章数量属性
+        for (Tag tag : tagList) {
+            Integer count = articleTagRefMapper.countArticleByTagId(tag.getTagId());
+            tag.setArticleCount(count);
+        }
+        return tagList;
     }
 }
